위경도 Double로 변환

+++ backend/src/main/java/com/codeboogie/comfortbackend/feeling/model/Feeling.java
@@ -27,11 +27,11 @@ public class Feeling {
     private Date publishDate; //작성한 날짜
     private Date updateDate; //수정한 날짜
     private String text; //작성한 글
-    private float xcoord; //글 작성 위치 x 좌표
-    private float ycoord; //글 작성 위치 y 좌표
+    private Double xcoord; //글 작성 위치 x 좌표
+    private Double ycoord; //글 작성 위치 y 좌표
 
     @Builder //빌더로 entity 만들어 한번에 저장
-    public Feeling(String id, Long userId, int score, Date publishDate, Date updateDate, String text, float xcoord, float ycoord) {
+    public Feeling(String id, Long userId, int score, Date publishDate, Date updateDate, String text, Double xcoord, Double ycoord) {
         this.id = id;
         this.userId = userId;
         this.score = score;
